GH-3212: Move FileSender to Fuseki test area
